+++ java/client/src/main/java/org/signal/libsignal/net/KeyTransparencyClient.java
@@ -169,7 +169,13 @@ public class KeyTransparencyClient {
   /**
    * Issue a monitor request to the key transparency service.
    *
-   * <p>Only the ACI is required to identify the account.
+   * <p>Store must contain data associated with the accoung being requested prior to making this
+   * call. Another way of putting this is: monitor cannot be called before {@link #search}.
+   *
+   * <p>If any of the monitored fields in the server response contain a version that is higher than
+   * the one currently in the store, a search request will be performed automatically and, if it
+   * succeeds, the updated account data will be stored. Otherwise, if the monitor does not detect
+   * any new versions, a search request will not be triggered.
    *
    * <p>If the latest distinguished tree head is not present in the store, it will be requested from
    * the server prior to performing the search via {@link #updateDistinguished}.
@@ -188,7 +194,13 @@ public class KeyTransparencyClient {
    * </ul>
    *
    * @param aci the ACI of the account to be searched for. Required.
+   * @param aciIdentityKey {@link IdentityKey} associated with the ACI. Required. Although key
+   *     transparency is used to request this information, passing it in the request lets chat
+   *     server know that it is not a random guess and that the caller of this API has communicated
+   *     with the ACI.
    * @param e164 string representation of an E.164 number associated with the account. Optional.
+   * @param unidentifiedAccessKey provides the same guess protections as the ACI identity key. This
+   *     parameter has the same optionality as the E.164 parameter.
    * @param usernameHash hash of the username associated with the account. Optional.
    * @param store local persistent storage for key transparency-related data, such as the latest
    *     tree heads and account monitoring data. It will be queried for data before performing the
@@ -201,23 +213,33 @@ public class KeyTransparencyClient {
    */
   public CompletableFuture<Void> monitor(
       /* @NotNull */ final ServiceId.Aci aci,
+      /* @NotNull */ final IdentityKey aciIdentityKey,
       final String e164,
+      final byte[] unidentifiedAccessKey,
       final byte[] usernameHash,
       final Store store) {
     Optional<byte[]> lastDistinguishedTreeHead = store.getLastDistinguishedTreeHead();
     if (lastDistinguishedTreeHead.isEmpty()) {
       return this.updateDistinguished(store)
-          .thenCompose((ignored) -> this.monitor(aci, e164, usernameHash, store));
+          .thenCompose(
+              (ignored) ->
+                  this.monitor(
+                      aci, aciIdentityKey, e164, unidentifiedAccessKey, usernameHash, store));
     }
     try (NativeHandleGuard tokioContextGuard = this.tokioAsyncContext.guard();
+        NativeHandleGuard identityKeyGuard = aciIdentityKey.getPublicKey().guard();
         NativeHandleGuard chatConnectionGuard = new NativeHandleGuard(chatConnection)) {
       return Native.KeyTransparency_Monitor(
               tokioContextGuard.nativeHandle(),
               this.environment.value,
               chatConnectionGuard.nativeHandle(),
               aci.toServiceIdFixedWidthBinary(),
+              identityKeyGuard.nativeHandle(),
               e164,
+              unidentifiedAccessKey,
               usernameHash,
+              // Technically this is a required parameter, but passing null
+              // to generate the error on the Rust side.
               store.getAccountData(aci).orElse(null),
               lastDistinguishedTreeHead.get())
           .thenApply(