+++ rust/bridge/shared/src/net/keytrans.rs
@@ -5,6 +5,7 @@
 
 use std::time::SystemTime;
 
+use itertools::Itertools;
 use libsignal_bridge_macros::{bridge_fn, bridge_io};
 use libsignal_bridge_types::net::chat::UnauthenticatedChatConnection;
 pub use libsignal_bridge_types::net::{Environment, TokioAsyncContext};
@@ -14,7 +15,7 @@ use libsignal_keytrans::{
     AccountData, KeyTransparency, LocalStateUpdate, StoredAccountData, StoredTreeHead,
 };
 use libsignal_net::keytrans::{
-    Error, Kt, KtApi as _, MaybePartial, SearchKey, SearchResult, UsernameHash,
+    monitor_and_search, Error, Kt, KtApi as _, MaybePartial, SearchKey, SearchResult, UsernameHash,
 };
 use libsignal_protocol::PublicKey;
 use prost::{DecodeError, Message};
@@ -106,21 +107,7 @@ async fn KeyTransparency_Search(
         config: Default::default(),
     };
 
-    let e164_pair = match (e164, unidentified_access_key) {
-        (None, None) => None,
-        (Some(e164), Some(uak)) => Some((e164, uak.into_vec())),
-        // technically harmless, but still invalid
-        (None, Some(_uak)) => {
-            return Err(Error::InvalidRequest(
-                "Unidentified access key without an E164",
-            ))
-        }
-        (Some(_e164), None) => {
-            return Err(Error::InvalidRequest(
-                "E164 without unidentified access key",
-            ))
-        }
-    };
+    let e164_pair = make_e164_pair(e164, unidentified_access_key)?;
 
     let account_data = account_data
         .map(|bytes| {
@@ -158,18 +145,27 @@ async fn KeyTransparency_Search(
 }
 
 #[bridge_io(TokioAsyncContext, node = false, ffi = false)]
+#[allow(clippy::too_many_arguments)]
 async fn KeyTransparency_Monitor(
     // TODO: it is currently possible to pass an env that does not match chat
     environment: AsType<Environment, u8>,
     chatConnection: &UnauthenticatedChatConnection,
     aci: Aci,
+    aci_identity_key: &PublicKey,
     e164: Option<E164>,
+    unidentified_access_key: Option<Box<[u8]>>,
     username_hash: Option<Box<[u8]>>,
-    account_data: Box<[u8]>,
+    // Bridging this as optional even though it is required because it is
+    // simpler to produce an error once here than on all platforms.
+    account_data: Option<Box<[u8]>>,
     last_distinguished_tree_head: Box<[u8]>,
 ) -> Result<Vec<u8>, Error> {
     let chat = chatConnection;
     let username_hash = username_hash.map(UsernameHash::from);
+
+    let Some(account_data) = account_data else {
+        return Err(Error::InvalidRequest("account data not found in store"));
+    };
 
     let account_data = {
         let stored: StoredAccountData = try_decode(account_data)?;
@@ -191,15 +187,29 @@ async fn KeyTransparency_Monitor(
         chat,
         config: Default::default(),
     };
-    let updated_account_data = kt
-        .monitor(
-            &aci,
-            e164,
-            username_hash,
-            account_data,
-            &last_distinguished_tree_head,
-        )
-        .await?;
+
+    let e164_pair = make_e164_pair(e164, unidentified_access_key)?;
+    let MaybePartial {
+        inner: updated_account_data,
+        missing_fields,
+    } = monitor_and_search(
+        &kt,
+        &aci,
+        aci_identity_key,
+        e164_pair,
+        username_hash,
+        account_data,
+        &last_distinguished_tree_head,
+    )
+    .await?;
+
+    if !missing_fields.is_empty() {
+        return Err(Error::InvalidResponse(format!(
+            "Missing fields: {}",
+            missing_fields.iter().join(", ")
+        )));
+    }
+
     Ok(StoredAccountData::from(updated_account_data).encode_to_vec())
 }
 
@@ -235,4 +245,21 @@ async fn KeyTransparency_Distinguished(
     let updated_distinguished = StoredTreeHead::from((tree_head, tree_root));
     let serialized = updated_distinguished.encode_to_vec();
     Ok(serialized)
+}
+
+#[cfg(feature = "jni")]
+fn make_e164_pair(
+    e164: Option<E164>,
+    unidentified_access_key: Option<Box<[u8]>>,
+) -> Result<Option<(E164, Vec<u8>)>, Error> {
+    match (e164, unidentified_access_key) {
+        (None, None) => Ok(None),
+        (Some(e164), Some(uak)) => Ok(Some((e164, uak.into_vec()))),
+        (None, Some(_uak)) => Err(Error::InvalidRequest(
+            "Unidentified access key without an E164",
+        )),
+        (Some(_e164), None) => Err(Error::InvalidRequest(
+            "E164 without unidentified access key",
+        )),
+    }
 }