+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
 v0.66.2
 
 - keytrans: Detect new versions of account data in monitor responses and invoke search
+- keytrans: Replace monitor with monitor_and_search on Android